move cache implementation into cache module out of CLI
